fix(oppstart): address self-review of financials layer before push
- Gini computation: cast through numeric so the baseline group's
  sum(rank * revenue) and (n+1)*total terms can't overflow bigint.
  Without this the refresh function hard-errors against the ~500k AS
  rows in production once backfill catches up.
- Drain candidate selection: replace the limit-200 newest-orgnr window
  with a full anti-join over the AI-flagged population (~thousands).
  Previously the drain got stuck the moment the 200 newest had
  fetch_state rows — it never advanced to older AI orgnrs until the
  180-day retry window kicked in.
- Sidebar / button / page header label: "Finansial data" was neither
  English nor Norwegian; rename to "Regnskap" to match the source name
  (Regnskapsregisteret) and the rest of the Bokmål nav.
- financials-growth.tsx: lift new Date() out of JSX into a lazy
  useState initialiser so render stays pure (mirrors scroller.tsx's
  heroStale pattern).

+++ supabase/migrations/0064_brreg_financials.sql
@@ -209,12 +209,16 @@ begin
     from src
   ),
   agg as (
+    -- Gini math runs in numeric throughout. The baseline group can hit
+    -- ~500k rows × ~1e12 NOK revenues — the inner sum(rank_asc * revenue)
+    -- and the (n+1)*total term both overflow bigint without explicit casts.
     select
       fiscal_year, is_ai_relevant,
       max(n) as n,
       max(group_total) as group_total,
-      (2 * sum(rank_asc * revenue) - (max(n) + 1) * max(group_total))::numeric
-        / nullif(max(n)::numeric * max(group_total), 0) as gini,
+      (2::numeric * sum(rank_asc::numeric * revenue::numeric)
+         - (max(n)::numeric + 1) * max(group_total)::numeric)
+        / nullif(max(n)::numeric * max(group_total)::numeric, 0) as gini,
       sum(revenue) filter (where rank_desc <= 10) as top10_sum,
       sum(revenue) filter (where rank_desc <= greatest(1, max(n) / 100)) as top1pct_sum
     from ranked
